requirements.txt
asgiref==3.2.10
Django==3.1.1
django-simple-history==2.11.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1








